+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Phone password auth type [#434](https://github.com/rokwire/core-building-block/issues/434)
 
+### Added
+- Service accounts [#306](https://github.com/rokwire/core-building-block/issues/306)
 
 ## [1.22.0] - 2022-04-02
 ### Added